Start moving to Next.js api routes
