+++ .gitignore
@@ -7,7 +7,7 @@ node_modules
 .next
 
 *.generated.*
-server/database/generated
+web/database/generated
 
 npm-debug.log*
 yarn-debug.log*
