+++ now.json
@@ -4,18 +4,9 @@
     {
       "src": "web/next.config.js",
       "use": "@now/next"
-    },
-    {
-      "src": "server/routes/**/*.ts",
-      "use": "@now/node",
-      "config": { "maxLambdaSize": "10mb" }
     }
   ],
   "routes": [
-    {
-      "src": "/api(.*)",
-      "dest": "/server/routes$1"
-    },
     {
       "src": "/(.*)",
       "dest": "/web/$1"
