+++ package.json
@@ -6,23 +6,21 @@
   "author": "Max Stoiber <contact@mxstbr.com>",
   "workspaces": {
     "packages": [
-      "web",
-      "server"
+      "web"
     ]
   },
   "scripts": {
     "prettier": "prettier",
     "prettify": "prettier --write",
     "lint": "eslint --ignore-path .gitignore --ext .js,.ts,.tsx web server",
-    "dev": "concurrently -n now,generate:web,generate:server,generate:db \"now dev\" \"yarn generate:web --watch\" \"nodemon -w server/graphql -x 'yarn generate:server'\" \"nodemon -w server/database/datamodel.prisma -x 'yarn run generate:db'\"",
+    "dev": "concurrently -n web,generate:web,generate:schema,generate:db \"yarn workspace web dev\" \"yarn generate:web --watch\" \"nodemon -w web/graphql -x 'yarn generate:schema'\" \"nodemon -w web/database/datamodel.prisma -x 'yarn run generate:db'\"",
     "web": "yarn workspace web",
-    "server": "yarn workspace server",
-    "db": "yarn workspace server run db",
-    "db:deploy": "yarn workspace server run db:deploy",
-    "generate": "yarn run generate:db && yarn run generate:server && yarn run generate:web",
+    "db": "yarn workspace web run db",
+    "db:deploy": "yarn workspace web run db:deploy",
+    "generate": "yarn run generate:db && yarn run generate:schema && yarn run generate:web",
     "generate:web": "graphql-codegen --config graphql-codegen.yml",
-    "generate:server": "ts-node --skip-project server/graphql/schema/index.ts",
-    "generate:db": "yarn workspace server generate:db",
+    "generate:schema": "ts-node --skip-project web/graphql/schema/index.ts",
+    "generate:db": "yarn workspace web generate:db",
     "postinstall": "yarn run generate"
   },
   "eslintConfig": {
@@ -63,7 +61,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged && yarn run generate:server && git add server/graphql/schema.graphql",
+      "pre-commit": "lint-staged && yarn run generate:schema && git add web/graphql/schema.graphql",
       "post-checkout": "yarn run generate && yarn run db:deploy"
     }
   },
