--- server/graphql/schema.graphql
@@ -1,19 +1,0 @@
-### This file was autogenerated by GraphQL Nexus
-### Do not make changes to this file directly
-
-
-type Query {
-  user(where: UserWhereUniqueInput!): User
-  viewer: User
-}
-
-type User {
-  avatarUrl: String
-  id: ID!
-  name: String!
-}
-
-input UserWhereUniqueInput {
-  googleId: ID
-  id: ID
-}
