--- server/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "server",
-  "private": true,
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "apollo-server-express": "^2.5.1",
-    "body-parser": "^1.19.0",
-    "cookie-parser": "^1.4.4",
-    "cookie-session": "^1.3.3",
-    "express": "^4.17.1",
-    "graphql": "^14.3.1",
-    "nexus": "^0.12.0-beta.6",
-    "nexus-prisma": "^0.3.7",
-    "passport": "^0.4.0",
-    "passport-google-oauth20": "^2.0.0",
-    "prisma": "^1.34.1"
-  },
-  "scripts": {
-    "db": "docker-compose up",
-    "db:deploy": "prisma deploy",
-    "generate:db": "prisma generate && nexus-prisma-generate --client ./database/generated/client --output ./database/generated/nexus-prisma"
-  },
-  "devDependencies": {
-    "@types/body-parser": "^1.17.0",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/cookie-session": "^2.0.37",
-    "@types/express": "^4.16.1",
-    "@types/graphql": "^14.2.0",
-    "@types/passport": "^1.0.0",
-    "@types/passport-google-oauth20": "^2.0.0",
-    "nexus-prisma-generate": "^0.3.7"
-  }
-}
