--- server/types.ts
@@ -1,6 +1,0 @@
-import { prisma } from "./database/generated/client";
-
-export interface Context {
-  prisma: typeof prisma;
-  viewerId?: string;
-}
