+++ web/package.json
@@ -11,7 +11,8 @@
     "@types/rebass": "^3.0.4",
     "@types/styled-components": "4.1.8",
     "@zeit/next-typescript": "^1.1.1",
-    "apollo-boost": "^0.3.1",
+    "apollo-boost": "^0.4.1",
+    "apollo-server-micro": "^2.6.9",
     "babel-plugin-styled-components": "^1.10.0",
     "fork-ts-checker-webpack-plugin": "^1.3.4",
     "graphql-tag": "^2.10.1",
@@ -23,7 +24,26 @@
     "react-dom": "^16.8.6",
     "rebass": "^3.1.1",
     "styled-components": "^4.2.0",
-    "styled-normalize": "^8.0.6"
+    "styled-normalize": "^8.0.6",
+    "apollo-server-express": "^2.5.1",
+    "body-parser": "^1.19.0",
+    "cookie-parser": "^1.4.4",
+    "cookie-session": "^1.3.3",
+    "express": "^4.17.1",
+    "graphql": "^14.3.1",
+    "nexus": "^0.12.0-beta.6",
+    "nexus-prisma": "^0.3.7",
+    "passport": "^0.4.0",
+    "passport-google-oauth20": "^2.0.0",
+    "prisma": "^1.34.1",
+    "@types/body-parser": "^1.17.0",
+    "@types/cookie-parser": "^1.4.1",
+    "@types/cookie-session": "^2.0.37",
+    "@types/express": "^4.16.1",
+    "@types/graphql": "^14.2.0",
+    "@types/passport": "^1.0.0",
+    "@types/passport-google-oauth20": "^2.0.0",
+    "nexus-prisma-generate": "^0.3.7"
   },
   "resolutions": {
     "@types/styled-components": "4.1.8",
@@ -31,9 +51,13 @@
   },
   "scripts": {
     "dev": "next",
-    "build": "next build"
+    "build": "next build",
+    "db": "docker-compose up",
+    "db:deploy": "prisma deploy",
+    "generate:db": "prisma generate && nexus-prisma-generate --client ./database/generated/client --output ./database/generated/nexus-prisma"
   },
   "devDependencies": {
-    "@types/rebass": "^3.0.4"
+    "@types/rebass": "^3.0.4",
+    "dotenv-webpack": "^1.7.0"
   }
 }
